+++ pages/myinfo/index.page.tsx
@@ -5,34 +5,20 @@ import Link from 'next/link';
 import { useRouter } from 'next/router';
 
 import { cn } from '@utils/mergeStyle';
-import { getRandomIndex } from '@utils/randomNumber';
 
 import { useDropdownStore } from '@stores/dropdownStore';
-import { useModalStore } from '@stores/modalStore';
 import { useUserInfoStore } from '@stores/userInfoStore';
 
 import QueryErrorBoundary from '@components/common/QueryErrorBoundary';
 import DevGuriError from '@components/common/error/DevGuriError';
-import {
-  MYINFO_NICKNAME_COMPLELTE_MODAL,
-  MYINFO_NICKNAME_EDIT_MODAL,
-  MYINFO_NICKNAME_RESULT_10_MODAL,
-  MYINFO_NICKNAME_RESULT_20_MODAL,
-  MYINFO_NICKNAME_RESULT_MODAL,
-} from '@components/common/modals/modalConfig/myInfoNickname';
 
-import {
-  NICKNAME_MODAL_FIRST_OVER_COUNT,
-  NICKNAME_MODAL_SECOND_OVER_COUNT,
-  NO_USER_NAME,
-} from '@/constants/UserInfoConstants';
+import { NO_USER_NAME } from '@/constants/UserInfoConstants';
 import { ROUTES } from '@/constants/routes';
 import { useMediaQueryContext } from '@/contexts/MediaQueryContext';
-import pencilIcon from '@/public/image/myInfo/pencil_level4.svg';
+import pencilIcon from '@/public/image/myInfo/pencil_level1.svg';
 import { UserInfoType } from '@/types/userInfoType';
 
-import NicknameComplete from './components/NicknameComplete';
-import NicknameResultModal from './components/NicknameResultModal';
+import { useNicknameModals } from './hooks/useNicknameModals';
 
 export default function MyInfo({ children }: { children: ReactNode }) {
   const router = useRouter();
@@ -40,7 +26,8 @@ export default function MyInfo({ children }: { children: ReactNode }) {
 
   const { setSort } = useDropdownStore();
   const { userInfo } = useUserInfoStore();
-  const { pushModal, popModal } = useModalStore();
+
+  const { handleNicknameEditClick } = useNicknameModals();
 
   const { isMobile } = useMediaQueryContext();
 
@@ -83,101 +70,6 @@ export default function MyInfo({ children }: { children: ReactNode }) {
       startHref: ROUTES.MY_INFO.ACCOUNT_DELETE,
     },
   ];
-
-  const pushNicknameResult20Modal = (count: number) => {
-    const nextCount = count + 1;
-
-    const randomTitles = [
-      '이제 진짜 마지막이에요… {nickname}, 더는 못 바꿔드려요!',
-      '하… 정말 끈질기시네… {nickname}, 이거면 됐죠?',
-      '이제 그만!! 후우... 이건 어때요? {nickname}',
-    ];
-
-    popModal();
-    pushModal({
-      ...MYINFO_NICKNAME_RESULT_20_MODAL,
-      contents: (
-        <NicknameResultModal
-          count={count}
-          title={randomTitles[getRandomIndex(3)]}
-          contents={MYINFO_NICKNAME_RESULT_20_MODAL.contents}
-        />
-      ),
-      submitFunction: () => pushCompleteModal(),
-      cancelFunction: () => pushNicknameResult20Modal(nextCount),
-    });
-  };
-
-  const pushNicknameResult10Modal = (count: number) => {
-    const nextCount = count + 1;
-
-    const randomTitles = [
-      '으음~ 조금 까다로우시네 {nickname} 이건 만족하시죠?',
-      '하아… 많이 고민했어요… {nickname} 이 정도 퀄리티면 인정 아닌가요?',
-      '허허… 쉽지 않았습니다… 하지만 {nickname}, 결국 이게 제일 잘 어울려요!',
-    ];
-
-    popModal();
-    pushModal({
-      ...MYINFO_NICKNAME_RESULT_10_MODAL,
-      contents: (
-        <NicknameResultModal
-          count={count}
-          title={randomTitles[getRandomIndex(3)]}
-          contents={MYINFO_NICKNAME_RESULT_10_MODAL.contents}
-        />
-      ),
-      submitFunction: () => pushCompleteModal(),
-      cancelFunction: () =>
-        count >= NICKNAME_MODAL_SECOND_OVER_COUNT - 1
-          ? pushNicknameResult20Modal(nextCount)
-          : pushNicknameResult10Modal(nextCount),
-    });
-  };
-
-  const pushNicknameResultModal = (count: number) => {
-    const nextCount = count + 1;
-
-    const randomTitles = [
-      '짜잔 ~ {nickname} 딱 어울려요!',
-      '이제부터 {nickname}으로 불릴 거예요, 기분 좋죠?',
-      '두근두근~ {nickname}, 정말 매력적인 이름이에요!',
-    ];
-
-    popModal();
-    pushModal({
-      ...MYINFO_NICKNAME_RESULT_MODAL,
-      contents: (
-        <NicknameResultModal
-          count={count}
-          title={randomTitles[getRandomIndex(3)]}
-          contents={MYINFO_NICKNAME_RESULT_MODAL.contents}
-        />
-      ),
-      submitFunction: () => pushCompleteModal(),
-      cancelFunction: () =>
-        count >= NICKNAME_MODAL_FIRST_OVER_COUNT - 1
-          ? pushNicknameResult10Modal(nextCount)
-          : pushNicknameResultModal(nextCount),
-    });
-  };
-
-  const pushCompleteModal = () => {
-    popModal();
-    pushModal({
-      ...MYINFO_NICKNAME_COMPLELTE_MODAL,
-      contents: <NicknameComplete />,
-      submitFunction: () => popModal(),
-    });
-  };
-
-  const handleNicknameEditClick = () => {
-    pushModal({
-      ...MYINFO_NICKNAME_EDIT_MODAL,
-      submitFunction: () => pushNicknameResultModal(1),
-      cancelFunction: () => popModal(),
-    });
-  };
 
   return (
     <div
